Refactor MCP Manager and Web Search Server: - introduce dedicated plugin-code-search server for plugin_code_search MCP tool (Freva plugin code retrieval) - cleanup of web-search server - new docker container for this tool, populated docker-compose.dev.yml + .env.example accordingly
