+++ .env.example
@@ -16,10 +16,11 @@ CLIMATECLAW_MONGODB_DATABASE_NAME="chatbot" # The name of the MongoDB database t
 CLIMATECLAW_MONGODB_COLLECTION_NAME="threads" # The name of the MongoDB collection to use for the storage of threads
 
 # MCP SETTINGS
-CLIMATECLAW_AVAILABLE_MCP_SERVERS=rag-server,code-server,web-search-server  # The list of servers to connect via backends client
+CLIMATECLAW_AVAILABLE_MCP_SERVERS=rag-server,code-server,web-search-server,plugin-code-search-server  # The list of servers to connect via backends client
 CLIMATECLAW_RAG_SERVER_URL="http://rag-server:8050"  # IMPORTANT: Server URLs must be given as 'CLIMATECLAW_{server_name}_SERVER_URL' for successful parsing
 CLIMATECLAW_CODE_SERVER_URL="http://code-server:8051"
 CLIMATECLAW_WEB_SEARCH_SERVER_URL="http://web-search-server:8052"
+CLIMATECLAW_PLUGIN_CODE_SEARCH_SERVER_URL="http://plugin-code-search-server:8053"
 CLIMATECLAW_MCP_REQUEST_TIMEOUT_SEC=600
 
 # MONGODB SETTINGS
