+++ docker/plugin-code-search-server/pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "plugin-code-search-server"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "fastapi>=0.116.1",
+    "fastmcp>=2.12.4",
+    "httpx>=0.28.1",
+    "uvicorn[standard]>=0.35.0",
+    "openai",
+    "requests>=2.31.0",
+]
