+++ src/climateclaw/tools/plugin_code_search/available_plugins.md
